Actualizacion de dependencias, checkeos android lint, correccion ciclo de vida para fragment mapa

+++ app/src/main/java/cl/figonzal/lastquakechile/core/utils/FirebaseMessage.kt
@@ -45,7 +45,7 @@ fun Context.notification(data: List<Any>, pendingIntent: PendingIntent) {
 
             //Notify
             (this@notification.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager).notify(
-                quake.quakeCode.toInt(),
+                quake.quakeCode,
                 build()
             )
         }
